+++ src/containerapp/azext_containerapp/_up_utils.py
@@ -979,18 +979,6 @@ def _get_registry_details(cmd, app: "ContainerApp", source):
     )
 
 
-def _validate_containerapp_name(name):
-    is_valid = True
-    is_valid = is_valid and name.lower() == name
-    is_valid = is_valid and len(name) <= MAXIMUM_CONTAINER_APP_NAME_LENGTH
-    is_valid = is_valid and '--' not in name
-    name = name.replace('-', '')
-    is_valid = is_valid and name.isalnum()
-    is_valid = is_valid and name[0].isalpha()
-    if not is_valid:
-        raise ValidationError(f"Invalid Container App name {name}. A name must consist of lower case alphanumeric characters or '-', start with a letter, end with an alphanumeric character, cannot have '--', and must be less than {MAXIMUM_CONTAINER_APP_NAME_LENGTH} characters.")
-
-
 # attempt to populate defaults for managed env, RG, ACR, etc
 def _set_up_defaults(
     cmd,